fix bug 1460707: fixed python 3 lint errors

+++ webapp-django/crashstats/crashstats/tests/test_views.py
@@ -764,7 +764,7 @@ class TestViews(BaseTestViews):
                 for version in ['20.0', '19.0']:
                     for build_type in ['beta', 'release']:
                         response['hits'].append({
-                            'adi_count': long(random.randint(100, 1000)),
+                            'adi_count': random.randint(100, 1000),
                             'build_type': build_type,
                             'date': start_date,
                             'version': version
@@ -1101,7 +1101,7 @@ class TestViews(BaseTestViews):
                 for version in ['18.0b1', '18.0b2', '19.0']:
                     for build_type in ['beta', 'release']:
                         response['hits'].append({
-                            'adi_count': long(random.randint(100, 1000)),
+                            'adi_count': random.randint(100, 1000),
                             'build_type': build_type,
                             'date': start_date,
                             'version': version
